Testy navigace
The methods we will be using like goRoot are currently on a new version of the NavController in the current version of the alpha. This isn’t to be confused with the existing NavController in Ionic 3. The new implementation of NavController may be renamed before the official release of Ionic 4